+++ library/runtime-api/src/commonMain/kotlin/io/matthewnelson/kmp/tor/runtime/api/config/builders/SocksFlagBuilder.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Matthew Nelson
+ * Copyright (c) 2024 Matthew Nelson
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,11 +15,12 @@
  **/
 @file:Suppress("FunctionName", "PropertyName")
 
-package io.matthewnelson.kmp.tor.runtime.api.config
+package io.matthewnelson.kmp.tor.runtime.api.config.builders
 
 import io.matthewnelson.kmp.tor.core.api.annotation.KmpTorDsl
 import io.matthewnelson.kmp.tor.runtime.api.ThisBlock
 import io.matthewnelson.kmp.tor.runtime.api.apply
+import io.matthewnelson.kmp.tor.runtime.api.config.TorConfig
 import kotlin.jvm.JvmField
 import kotlin.jvm.JvmSynthetic
 
@@ -72,6 +73,7 @@ public class SocksFlagBuilder private constructor() {
             block: ThisBlock<SocksFlagBuilder>,
         ) {
             val b = SocksFlagBuilder().apply(block)
+
             b.NoIPv4Traffic?.let {
                 val flag = "NoIPv4Traffic"
                 if (it) flags.add(flag) else flags.remove(flag)
